+++ docs/changes.rst
@@ -4,6 +4,7 @@ Weblate 4.0
 Not yet released.
 
 * Weblate now requires Python 3.5 or newer.
+* Added management overview of component alerts.
 
 Weblate 3.11.1
 --------------